+++ src/Fable.Compiler/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 4.0.0-alpha-002 - 2024-01-25
+
 ### Changed
 
 * Respect file extension from `CliArgs`